litchi_backend/requirements.txt
amqp==2.5.2
asn1crypto==0.23.0
billiard==3.6.2.0
boto3==1.12.13
botocore==1.15.13
celery==4.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
defusedxml==0.6.0
Django==1.11.20
django-admin-view-permission==1.6
django-allauth==0.39.1
django-appconf==1.0.2
django-auditlog==0.4.5
django-axes==3.0.3
django-cors-headers==2.1.0
django-fernet-fields==0.5
django-filter==1.1.0
django-jsonfield==1.0.1
django-model-utils==3.1.2
django-storages==1.9.1
django-redis==4.11.0
django-redis-sessions==0.6.1
django-rest-auth==0.9.3
django-rest-swagger==2.1.2
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
drf-openapi==1.0.0
elastic-apm==6.1.3
idna==2.8
importlib-metadata==1.2.0
itypes==1.1.0
Jinja2==2.10.3
kombu==4.6.7
MarkupSafe==1.1.1
more-itertools==8.0.2
mysqlclient==1.3.13
numpy==1.17.4
oauthlib==3.0.1
openapi-codec==1.3.2
pandas==0.24.2
Pillow==7.2.0
pycparser==2.18
pyfcm==1.4.3
PyJWT==1.5.3
pyOpenSSL==18.0.0
PyPDF2==1.26.0
python-dateutil==2.6.0
python3-openid==3.1.0
pytz==2018.5
pycryptodome==3.15.0
pycrypto==2.6.1
redis==3.4.1
reportlab==3.5.47
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
shortuuid==1.0.1
simplejson==3.17.0
six==1.13.0
uritemplate==3.0.0
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
